+++ src/main/java/org/y1000/entities/players/kungfu/UnnamedBufa.java
@@ -2,7 +2,7 @@ package org.y1000.entities.players.kungfu;
 
 public class UnnamedBufa extends AbstractFootKungFu {
 
-    public UnnamedBufa(float level) {
+    public UnnamedBufa(int level) {
         super("无名步法", level);
     }
 }
